Added some functionalities to the refman to be able to go into other documents includes by <xs:include>. Also added support for unions of simple types (<xs:union>). With these changes the ERSL schema can be now displayed without making a flat version. Also made some small changes to the ERSL annotations: - changed img/href to img/src -> now the initial image will be displayed - added id "SimulationScenarion_ERSL" to the element definition of SimulationScenario and added a reference to it in the annotation.
ATTENTION!
Be aware that this change just applies partially to the AORSL schema because in AORLS <xs:redefine>, which is still not supported. Therefore for the AORSL schema a flat version is still needed.
Be also aware that theses changes were just tested by watching the transformation's result with ERSL. There might be bugs which were not discovered yet with features that are currently not used by the ERSL schema.